refactor(AYC-000): resolve frontend dependency-cruiser errors
- Fix circular dependency in features/theme by importing directly instead of barrel
- Fix widgets-no-pages: chart widgets import ToolUseMessageContentResponseDto from generated client instead of pages layer
- Remove resolved violations from known violations file
