+++ ayunis-core-frontend/.dependency-cruiser-known-violations.json
@@ -1,57 +1,5 @@
 [
   {
-    "type": "cycle",
-    "from": "src/features/theme/index.ts",
-    "to": "src/features/theme/useTheme.ts",
-    "rule": {
-      "severity": "error",
-      "name": "no-circular-dependencies"
-    },
-    "cycle": [
-      {
-        "name": "src/features/theme/useTheme.ts",
-        "dependencyTypes": [
-          "local",
-          "export"
-        ]
-      },
-      {
-        "name": "src/features/theme/index.ts",
-        "dependencyTypes": [
-          "local",
-          "import"
-        ]
-      }
-    ]
-  },
-  {
-    "type": "dependency",
-    "from": "src/widgets/charts/ui/BarChartWidget.tsx",
-    "to": "src/pages/chat/model/openapi.ts",
-    "rule": {
-      "severity": "error",
-      "name": "widgets-no-pages"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "src/widgets/charts/ui/LineChartWidget.tsx",
-    "to": "src/pages/chat/model/openapi.ts",
-    "rule": {
-      "severity": "error",
-      "name": "widgets-no-pages"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "src/widgets/charts/ui/PieChartWidget.tsx",
-    "to": "src/pages/chat/model/openapi.ts",
-    "rule": {
-      "severity": "error",
-      "name": "widgets-no-pages"
-    }
-  },
-  {
     "type": "dependency",
     "from": "src/widgets/app-sidebar/ui/ChatsSidebarGroup.tsx",
     "to": "src/widgets/confirmation-modal/index.ts",
